Refactor: type 파일 이름 변경 및 index 뽑기
