--- src/types/userInfo.ts
@@ -1,6 +1,0 @@
-export interface userInfo {
-  id: number;
-  user_name: string;
-  user_email: string;
-  phone_number: string;
-}
